scripts/requirements.txt
# Python requirements for ABC tools
pyyaml>=6.0
